+++ systemd/daemon.go
@@ -1,7 +1,7 @@
 package systemd
 
 import (
-	"github.com/coreos/go-systemd/dbus"
+	"github.com/coreos/go-systemd/v22/dbus"
 	"github.com/pkg/errors"
 )