+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [v0.2.1](https://github.com/HQJaTu/ct2stimer/releases/tag/v0.2.1) (2026-06-14)
+
+- Build on GoLang 1.26
+- Without any input, fail with error
+  - Don't default to `/etc/crontab`, its format differs
+
 # [v0.2.0](https://github.com/dtan4/ct2stimer/releases/tag/v0.2.0) (2017-02-23)
 
 ## Features